On branch v2 Your branch is up to date with 'origin/v2'.
Changes to be committed:
	modified:   app/Http/Controllers/eskulController.php
	deleted:    public/images/logo-eskul/paskibra.png
	modified:   resources/views/admin/eskul.blade.php

+++ app/Http/Controllers/eskulController.php
@@ -14,139 +14,31 @@ use App\Models\berita;
 class eskulController extends Controller
 {
     public function index(Request $request){
-        
-        // $data = eskul::latest();
 
-        // if(request('search')) {
-        //     $data->where('nama_eskul','slug' , '%' .request('search'). '%');
-        // }else{
-        //     $data = eskul::paginate(5);
+        $eskul = null;
+        $data_eskul = eskul::all();
+        $currentRole = role::where("role", "=", auth()->user()->role)->first();
 
-        // }
+        if ($request->has('search')) {
+            if (auth()->user()->role == "root") {
+                $eskul = eskul::where('nama_eskul', 'like', '%' . $request->search . '%')
+                    ->orWhere('id', 'like', '%' . $request->search . '%')->paginate(5);
+            } else {
+                $eskul = eskul::where('id_eskul', '=', $currentRole['eskul_id'])->where('nama_eskul', 'like', '%' . $request->search . '%')
+                    ->orWhere('id', 'like', '%' . $request->search . '%')->paginate(5);
+            }
+        } else {
+            if (auth()->user()->role == "root") {
+                $eskul = eskul::paginate(2);
+            } else {
 
-        if($request->has('search')){
-                // $adminitc = DB::
-
-            $adminmpk = DB::select('select * from eskuls where id = 1 ');
-            $adminosis = DB::select('select * from eskuls where id = 2 ');
-            $adminirma = DB::select('select * from eskuls where id = 3 ');
-            $adminpks = DB::select('select * from eskuls where id = 4 ');
-            $adminpramukaputra = DB::select('select * from eskuls where id = 5 ');
-            $adminpramukaputri = DB::select('select * from eskuls where id = 6 ');
-            $adminpaskibra = DB::select('select * from eskuls where id = 7 ');
-            $adminvolly = DB::select('select * from eskuls where id = 8 ');
-            $adminfutsalputra = DB::select('select * from eskuls where id = 9 ');
-            $adminfutsalputri = DB::select('select * from eskuls where id = 10 ');
-            $adminsepakbola = DB::select('select * from eskuls where id = 11 ');
-            $adminbasket = DB::select('select * from eskuls where id = 12 ');
-            $adminhockey = DB::select('select * from eskuls where id = 13 ');
-            $adminbadminton = DB::select('select * from eskuls where id = 14 ');
-            $adminkarate = DB::select('select * from eskuls where id = 15 ');
-            $admintaekwondo = DB::select('select * from eskuls where id = 16 ');
-            $adminsilat = DB::select('select * from eskuls where id = 17 ');
-            $admintarungderajat = DB::select('select * from eskuls where id = 18 ');
-            $adminkir = DB::select('select * from eskuls where id = 19 ');
-            $adminkopsis = DB::select('select * from eskuls where id = 20 ');
-            $adminpmr = DB::select('select * from eskuls where id = 21 ');
-            $adminsenitari = DB::select('select * from eskuls where id = 22 ');
-            $adminmarchingband = DB::select('select * from eskuls where id = 23 ');
-            $adminsenikriya = DB::select('select * from eskuls where id = 24 ');
-            $adminpaduansuara = DB::select('select * from eskuls where id = 25 ');
-            $adminkarawitan = DB::select('select * from eskuls where id = 26 ');
-            $adminteater = DB::select('select * from eskuls where id = 27 ');
-            $adminsenimusik = DB::select('select * from eskuls where id = 28 ');
-            $adminitclub = DB::select('select * from eskuls where id = 29 ');
-            $adminbroadcast = DB::select('select * from eskuls where id = 30 ');
-            $adminenglishclub = DB::select('select * from eskuls where id = 31 ');
-            $adminkoreanclub = DB::select('select * from eskuls where id = 32 ');
-            $adminjapanclub = DB::select('select * from eskuls where id = 33 ');
-            $admindeutschclub = DB::select('select * from eskuls where id = 34 ');
-            
-
-            $data = eskul::where('slug','like', '%' .$request->search. '%')->
-            orWhere('nama_eskul','like', '%' .$request->search. '%')->paginate(2);
-        }else{
-            
-            $adminmpk = DB::select('select * from eskuls where id = 1 ');
-            $adminosis = DB::select('select * from eskuls where id = 2 ');
-            $adminirma = DB::select('select * from eskuls where id = 3 ');
-            $adminpks = DB::select('select * from eskuls where id = 4 ');
-            $adminpramukaputra = DB::select('select * from eskuls where id = 5 ');
-            $adminpramukaputri = DB::select('select * from eskuls where id = 6 ');
-            $adminpaskibra = DB::select('select * from eskuls where id = 7 ');
-            $adminvolly = DB::select('select * from eskuls where id = 8 ');
-            $adminfutsalputra = DB::select('select * from eskuls where id = 9 ');
-            $adminfutsalputri = DB::select('select * from eskuls where id = 10 ');
-            $adminsepakbola = DB::select('select * from eskuls where id = 11 ');
-            $adminbasket = DB::select('select * from eskuls where id = 12 ');
-            $adminhockey = DB::select('select * from eskuls where id = 13 ');
-            $adminbadminton = DB::select('select * from eskuls where id = 14 ');
-            $adminkarate = DB::select('select * from eskuls where id = 15 ');
-            $admintaekwondo = DB::select('select * from eskuls where id = 16 ');
-            $adminsilat = DB::select('select * from eskuls where id = 17 ');
-            $admintarungderajat = DB::select('select * from eskuls where id = 18 ');
-            $adminkir = DB::select('select * from eskuls where id = 19 ');
-            $adminkopsis = DB::select('select * from eskuls where id = 20 ');
-            $adminpmr = DB::select('select * from eskuls where id = 21 ');
-            $adminsenitari = DB::select('select * from eskuls where id = 22 ');
-            $adminmarchingband = DB::select('select * from eskuls where id = 23 ');
-            $adminsenikriya = DB::select('select * from eskuls where id = 24 ');
-            $adminpaduansuara = DB::select('select * from eskuls where id = 25 ');
-            $adminkarawitan = DB::select('select * from eskuls where id = 26 ');
-            $adminteater = DB::select('select * from eskuls where id = 27 ');
-            $adminsenimusik = DB::select('select * from eskuls where id = 28 ');
-            $adminitclub = DB::select('select * from eskuls where id = 29 ');
-            $adminbroadcast = DB::select('select * from eskuls where id = 30 ');
-            $adminenglishclub = DB::select('select * from eskuls where id = 31 ');
-            $adminkoreanclub = DB::select('select * from eskuls where id = 32 ');
-            $adminjapanclub = DB::select('select * from eskuls where id = 33 ');
-            $admindeutschclub = DB::select('select * from eskuls where id = 34 ');
-
-            
-            
-            $data = eskul::paginate(2);
-
+                $eskul = eskul::where('id', '=', $currentRole['eskul_id'])->get();
+            }
         }
-        $role = role::all();
-
-        // dd($data);
+        
+        
         return view('admin.eskul', compact(
-            'data',
-            'role',
-            'adminmpk',
-            'adminosis',
-            'adminirma',
-            'adminpks',
-            'adminpramukaputra',
-            'adminpramukaputri',
-            'adminpaskibra',
-            'adminvolly',
-            'adminfutsalputra',
-            'adminfutsalputri',
-            'adminsepakbola',
-            'adminbasket',
-            'adminhockey',
-            'adminbadminton',
-            'adminkarate',
-            'admintaekwondo',
-            'adminsilat',
-            'admintarungderajat',
-            'adminkir',
-            'adminkopsis',
-            'adminpmr',
-            'adminsenitari',
-            'adminmarchingband',
-            'adminsenikriya',
-            'adminpaduansuara',
-            'adminkarawitan',
-            'adminteater',
-            'adminsenimusik',
-            'adminitclub',
-            'adminbroadcast',
-            'adminenglishclub',
-            'adminkoreanclub',
-            'adminjapanclub',
-            'admindeutschclub',
+            'eskul',
         )) ;
     }
     
@@ -208,6 +100,9 @@ class eskulController extends Controller
 
     public function deleteeskul($id){
         $data = eskul::find($id);
+        if (auth()->user()->role != 'root') {
+            return redirect()->route('eskul')->with('error', ' Data Gagal Di Delete');
+        }
         if(File_exists(public_path('images/logo-eskul/'.$data->logo))){ //either you can use file path instead of $data->image
             unlink(public_path('images/logo-eskul/'.$data->logo));//here you can also use path like as ('uploads/media/welcome/'. $data->image)
         }
